requirements.txt
psycopg2-binary==2.9.5
faker==18.2.1
python-dotenv==1.0.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Paginate==0.5.1
werkzeug==2.2.2
requests ==2.28.1
apache-airflow[postgres]== 2.7.1
pendulum==2.0.5